i forgot i already had refactored strains
